Refine JobManager
