+++ FhirToDataLake/.editorconfig
@@ -1,11 +1,17 @@
 ﻿[*.cs]
 
-# SA1600: Elements should be documented
-dotnet_diagnostic.SA1600.severity = none
-
 # SA1101: Prefix local calls with this
 dotnet_diagnostic.SA1101.severity = none
 
 # SA1309: Field names should not begin with underscore
 dotnet_diagnostic.SA1309.severity = none
 dotnet_diagnostic.SA0001.severity = none
+
+# SA1600: Elements should be documented
+dotnet_diagnostic.SA1600.severity = none
+
+# SA1623: The property's documentation summary text should begin with: 'Gets or sets'
+dotnet_diagnostic.SA1623.severity = none
+
+# SA1623: Documentation text should end with a period
+dotnet_diagnostic.SA1629.severity = none
