+++ FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Core/PipelineRegistrationExtensions.cs
@@ -17,6 +17,8 @@ namespace Microsoft.Health.Fhir.Synapse.Core
         {
             services.AddSingleton<IJobStore, AzureBlobJobStore>();
 
+            services.AddSingleton<JobManager, JobManager>();
+
             services.AddSingleton<JobExecutor, JobExecutor>();
 
             services.AddSingleton<ITaskExecutor, TaskExecutor>();
